move client writer that does the write out to clients into own source
